% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpta_predict.R
\name{fpta_predict}
\alias{fpta_predict}
\title{Predict the payout function f(x,y) with first person's trait = x and second person's trait as y}
\usage{
fpta_predict(list_basis = NULL, Mat_Embed_Coef = NULL, X1 = NULL, X2 = NULL)
}
\arguments{
\item{list_basis}{a list of k vectorized function provided by the user, the basis function should be the same as used in solve_embedding()}

\item{Mat_Embed_Coef}{2p x k coefficient matrix, output from solve_embedding()}

\item{X1}{n x p matrix, containing the traits of first n players}

\item{X2}{n x p matrix, containing the traits of second n players}
}
\value{
y, vector of length n, which is the estimated f(x,y)
}
\description{
Predict the payout function f(x,y) with first person's trait = x and second person's trait as y
}
\examples{
a = solve_embedding(f_val, list_f, c(0,1,1.5,2))
fpta_predict(list_f, a$Mat_Embed_Coef, rnorm(100), rnorm(100))
}
