% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_embedding.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot Embedding}
\usage{
plot_embedding(list_basis = NULL, Mat_Embed_Coef = NULL, X = NULL, i = 1)
}
\arguments{
\item{list_basis}{a list of k vectorized function provided by the user, the basis function should be the same as used in solve_embedding()}

\item{Mat_Embed_Coef}{2p x k coefficient matrix, output from solve_embedding()}

\item{X}{n x p matrix, containing the traits of n players}

\item{i}{integer, 1<= i <= p. Generate the plot for the i th embedding}
}
\value{
a plot
}
\description{
This function plots the embedding for the chosen samples for better visualization.
The payout of two agents can be understand as the area of the triangle formed by the two points and the origin.
}
\examples{
plot_embedding(list_basis, Mat_Embed_Coef, X, 1)

}
