% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_embedding.R
\name{update_embedding}
\alias{update_embedding}
\title{This calculates the disc game embedding of the coefficients associated with the basis function provided by the user}
\usage{
update_embedding(projection, GramMatrix)
}
\arguments{
\item{projection}{k x k projection matrix generated by update_projection}

\item{GramMatrix}{the matrix generated by Gram-Schmit process}
}
\value{
a list containing a vector lambda, which calculates the relative importance of the embedding, an 2p x k coefficient matrix, Mat_Embed_Coef
where 2i-1 and 2i row is the coefficients of the basis function associated with the ith embedding. The 2i-1 row is the embedding for the x-coordinate, the 2i th row is the embedding for the y-coordinate.
}
\description{
This calculates the disc game embedding of the coefficients associated with the basis function provided by the user
}
