% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_projection.R
\name{update_projection}
\alias{update_projection}
\title{Calculate the projection matrix based on Hilbert projection theorem}
\usage{
update_projection(B, f_val, GramMatrix)
}
\arguments{
\item{B}{n x k matrix, with B_ij represented by ith data point evaluated using j th basis function}

\item{f_val}{n x n  matrix, with i,j th entry representing f(x_i, x_j)}

\item{GramMatrix}{the matrix generated by Gram-Schmit process}
}
\value{
a k x k projection matrix
}
\description{
Calculate the projection matrix based on Hilbert projection theorem
}
